NameEntity is a sealed class
I think it's safe to assume that this API won't change and this actually simplifies a lot
of processing stuff (which one cane see in nameEntityProcessing)
