--- ast-common/src/org/jetbrains/dukat/astCommon/IdentifierEntity.kt
@@ -1,5 +1,0 @@
-package org.jetbrains.dukat.astCommon
-
-data class IdentifierEntity(
-        val value: String
-) : NameEntity
