+++ ast-common/src/org/jetbrains/dukat/astCommon/NameEntity.kt
@@ -1,3 +1,12 @@
 package org.jetbrains.dukat.astCommon
 
-interface NameEntity : Entity
+sealed class NameEntity : Entity
+
+data class IdentifierEntity(
+        val value: String
+) : NameEntity()
+
+data class QualifierEntity(
+        val left: NameEntity,
+        val right: IdentifierEntity
+) : NameEntity()
