--- ast-common/src/org/jetbrains/dukat/astCommon/QualifierEntity.kt
@@ -1,6 +1,0 @@
-package org.jetbrains.dukat.astCommon
-
-data class QualifierEntity(
-        val left: NameEntity,
-        val right: IdentifierEntity
-) : NameEntity
