+++ ast/src/org/jetbrains/dukat/ast/model/nodes/processing/nameEntityProcessing.kt
@@ -3,13 +3,11 @@ package org.jetbrains.dukat.ast.model.nodes.processing
 import org.jetbrains.dukat.astCommon.IdentifierEntity
 import org.jetbrains.dukat.astCommon.NameEntity
 import org.jetbrains.dukat.astCommon.QualifierEntity
-import org.jetbrains.dukat.panic.raiseConcern
 
 fun NameEntity.toNode(): NameEntity {
     return when (this) {
         is IdentifierEntity -> IdentifierEntity(value)
         is QualifierEntity -> QualifierEntity(left = left.toNode(), right = IdentifierEntity(right.value))
-        else -> raiseConcern("unknown QualifiedLeftDeclaration") { IdentifierEntity(this.toString()) }
     }
 }
 
@@ -17,7 +15,6 @@ private fun NameEntity.countDepth(current: Int): Int {
     return when (this) {
         is IdentifierEntity -> current + 1
         is QualifierEntity -> left.countDepth(current) + right.countDepth(current)
-        else -> raiseConcern("unknown org.jetbrains.dukat.astCommon.NameEntity ${this}") { 0 }
     }
 }
 
@@ -25,7 +22,6 @@ fun NameEntity.process(handler: (String) -> String): NameEntity {
     return when (this) {
         is IdentifierEntity -> IdentifierEntity(handler(value))
         is QualifierEntity -> copy(left = left.process(handler), right = right.process(handler) as IdentifierEntity)
-        else -> raiseConcern("failed to process NameEntity ${this}") { this }
     }
 }
 
@@ -40,7 +36,6 @@ private fun IdentifierEntity.appendLeft(qualifiedLeftNode: NameEntity): Qualifie
     return when (qualifiedLeftNode) {
         is IdentifierEntity -> this.appendLeft(qualifiedLeftNode)
         is QualifierEntity -> this.appendLeft(qualifiedLeftNode)
-        else -> raiseConcern("unknown NameEntity ${qualifiedLeftNode}") { QualifierEntity(this, this) }
     }
 }
 
@@ -48,7 +43,6 @@ private fun QualifierEntity.appendLeft(qualifiedNode: NameEntity): QualifierEnti
     return when (left) {
         is IdentifierEntity -> copy(left = (left as IdentifierEntity).appendLeft(qualifiedNode))
         is QualifierEntity -> copy(left = (left as QualifierEntity).appendLeft(qualifiedNode))
-        else -> raiseConcern("unknown NameEntity ${left}") { this }
     }
 }
 
@@ -56,7 +50,6 @@ fun NameEntity.appendLeft(qualifiedNode: NameEntity): NameEntity {
     return when (this) {
         is IdentifierEntity -> this.appendLeft(qualifiedNode)
         is QualifierEntity -> this.appendLeft(qualifiedNode)
-        else -> raiseConcern("unknown NameEntity ${this}") { this }
     }
 }
 
@@ -73,10 +66,8 @@ fun IdentifierEntity.appendLeft(qualifiedNode: QualifierEntity): QualifierEntity
             QualifierEntity(when (left) {
                 is IdentifierEntity -> this.appendLeft(left)
                 is QualifierEntity -> this.appendLeft(left)
-                else -> raiseConcern("unkown qualifiedNode $left") { qualifiedNode }
             }, nodeLeft.right)
         }
-        else -> raiseConcern("unkown qualifiedNode $nodeLeft") { qualifiedNode }
     }
     return QualifierEntity(left, qualifiedNode.right)
 }
@@ -85,7 +76,6 @@ fun IdentifierEntity.appendRight(qualifiedLeftNode: NameEntity): QualifierEntity
     return when (qualifiedLeftNode) {
         is IdentifierEntity -> this.appendRight(qualifiedLeftNode)
         is QualifierEntity -> this.appendRight(qualifiedLeftNode)
-        else -> raiseConcern("unknown NameEntity ${qualifiedLeftNode}") { QualifierEntity(qualifiedLeftNode, this) }
     }
 }
 
@@ -106,7 +96,6 @@ fun QualifierEntity.appendRight(qualifiedNode: QualifierEntity): QualifierEntity
     return when (nodeLeft) {
         is IdentifierEntity -> appendRight(nodeLeft).appendRight(qualifiedNode.right)
         is QualifierEntity -> appendRight(nodeLeft).appendRight(qualifiedNode.right)
-        else -> raiseConcern("unknown QualifierEntity") { this }
     }
 }
 
@@ -115,14 +104,11 @@ fun NameEntity.appendRight(qualifiedNode: NameEntity): NameEntity {
         is IdentifierEntity -> when (qualifiedNode) {
             is IdentifierEntity -> appendRight(qualifiedNode)
             is QualifierEntity -> appendRight(qualifiedNode)
-            else -> raiseConcern("unknown QualifierEntity") { this }
         }
         is QualifierEntity -> when (qualifiedNode) {
             is IdentifierEntity -> appendRight(qualifiedNode)
             is QualifierEntity -> appendRight(qualifiedNode)
-            else -> raiseConcern("unknown QualifierEntity") { this }
         }
-        else -> raiseConcern("unknown QualifierEntity") { this }
     }
 }
 
@@ -130,7 +116,6 @@ fun NameEntity.shiftRight(): NameEntity? {
     return when (this) {
         is IdentifierEntity -> null
         is QualifierEntity -> left
-        else -> raiseConcern("unknown NameEntity") { this }
     }
 }
 
@@ -138,7 +123,6 @@ fun NameEntity.leftMost(): NameEntity {
     return when (this) {
         is IdentifierEntity -> this
         is QualifierEntity -> left.leftMost()
-        else -> raiseConcern("unknown NameEntity ${this}") { this }
     }
 }
 
@@ -146,7 +130,6 @@ fun NameEntity.rightMost(): NameEntity {
     return when (this) {
         is IdentifierEntity -> this
         is QualifierEntity -> right
-        else -> raiseConcern("unknown NameEntity ${this}") { this }
     }
 }
 
@@ -161,7 +144,6 @@ fun NameEntity.shiftLeft(): NameEntity? {
                 QualifierEntity(leftShifted, right)
             }
         }
-        else -> raiseConcern("unknown NameEntity") { this }
     }
 }
